src/requirements.txt
aioredis==1.2.0
asgiref==3.1.2
async-timeout==3.0.1
attrs==19.1.0
autobahn==19.5.1
Automat==0.7.0
channels==2.2.0
channels-redis==2.4.0
constantly==15.1.0
daphne==2.3.0
Django==2.2.1
django-cors-headers==3.0.1
django-crispy-forms==1.7.2
djangorestframework==3.9.4
gunicorn==19.9.0
hiredis==1.0.0
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
msgpack==0.6.1
psycopg2==2.8.2
PyHamcrest==1.9.0
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
Twisted==19.2.0
txaio==18.8.1
zope.interface==4.6.0
